+++ packages/openworkflow/postgres/backend.ts
@@ -1,11 +1,4 @@
 import {
-  newPostgres,
-  newPostgresMaxOne,
-  Postgres,
-  migrate,
-  DEFAULT_SCHEMA,
-} from "./postgres.js";
-import {
   DEFAULT_NAMESPACE_ID,
   Backend,
   CancelWorkflowRunParams,
@@ -23,11 +16,18 @@ import {
   FailWorkflowRunParams,
   CompleteWorkflowRunParams,
   SleepWorkflowRunParams,
-  StepAttempt,
-  WorkflowRun,
-  DEFAULT_RETRY_POLICY,
-  JsonValue,
-} from "openworkflow/internal";
+} from "../backend.js";
+import { JsonValue } from "../core/json.js";
+import { DEFAULT_RETRY_POLICY } from "../core/retry.js";
+import { StepAttempt } from "../core/step.js";
+import { WorkflowRun } from "../core/workflow.js";
+import {
+  newPostgres,
+  newPostgresMaxOne,
+  Postgres,
+  migrate,
+  DEFAULT_SCHEMA,
+} from "./postgres.js";
 
 const DEFAULT_PAGINATION_PAGE_SIZE = 100;